aura_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.40
langchain-core==0.1.52
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pillow==10.1.0
aiofiles==23.2.1
typing-extensions==4.8.0
